Moved HP tests into their own files
